+++ js/tamarin/core/MathUtils.cpp
@@ -1418,7 +1418,7 @@ namespace avmplus
 
 		/* The sequence always starts with 1. */
 		//    pRandomFast->uValue = 1L;
-		pRandomFast->uValue = OSDep::currentTimeMillis();
+		pRandomFast->uValue = (uint32)(OSDep::currentTimeMillis());
 
 		/* Figure out the sequence length (2^n - 1). */
 		pRandomFast->uSequenceLength = (1L << n) - 1L;
