+++ js/tamarin/core/Multiname.cpp
@@ -125,7 +125,7 @@ namespace avmplus
 		// find a matching namespace between m2 and this
 		// if no match return false
 		Stringp u2 = qname->getNamespace()->getURI();
-        uint8 type2 = qname->getNamespace()->getType();
+        uint8 type2 = (uint8)(qname->getNamespace()->getType());
 		//Stringp s2 = core->string(u2);
 		for (int i = 0; i < this->namespaceCount(); i++)
 		{
@@ -133,7 +133,7 @@ namespace avmplus
 			// to check URIs since two namespaces with different prefixes
 			// are considered a match
 			Stringp u1 = getNamespace(i)->getURI();
-            uint8 type1 = getNamespace(i)->getType();
+            uint8 type1 = (uint8)(getNamespace(i)->getType());
 			//Stringp s1 = core->string(u1);
 			if (u1 == u2 && type1 == type2)
 				return true;
